+++ modules/datacafe-client/src/main/java/edu/emory/bmi/datacafe/client/query_builder/package-info.java
@@ -1,0 +1,4 @@
+/*
+ * Building queries for Data Cafe client.
+ */
+package edu.emory.bmi.datacafe.client.query_builder;
