+++ modules/datacafe-core/src/main/java/edu/emory/bmi/datacafe/core/hazelcast/config/HzConstants.java
@@ -38,5 +38,4 @@ public final class HzConstants {
     public static final int LAST = HAZELCAST_INSTANCES_STARTED_SIMULTANEOUSLY - 1;
 
     public static final String MAIN_HZ_CLUSTER = HzConfigReader.getMainClusterName();
-    public static final String SUB_HZ_CLUSTER = HzConfigReader.getSubClusterName();
 }
