+++ modules/datacafe-server/src/main/java/edu/emory/bmi/datacafe/mongo/ext/DBCursorConnector.java
@@ -1,31 +1,35 @@
 /*
- * Copyright (c) 2015-2016, Pradeeban Kathiravelu and others. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package edu.emory.bmi.datacafe.mongo;
+* Copyright (c) 2015-2016, Pradeeban Kathiravelu and others. All Rights Reserved.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+* http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+package edu.emory.bmi.datacafe.mongo.ext;
 
 import com.mongodb.BasicDBObject;
 import com.mongodb.DBCollection;
 import com.mongodb.DBCursor;
 import edu.emory.bmi.datacafe.constants.MongoConstants;
+import edu.emory.bmi.datacafe.mongo.MongoCollection;
+import edu.emory.bmi.datacafe.mongo.MongoConnector;
+import edu.emory.bmi.datacafe.mongo.MongoUtil;
 
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Connects through the DBCursor methods.
- */
+* Connects through the DBCursor methods.
+*/
+@Deprecated
 public class DBCursorConnector extends MongoConnector {
 
     /**
