+++ modules/datacafe-server/src/main/java/edu/emory/bmi/datacafe/mongo/ext/FindIterableConnector.java
@@ -1,30 +1,33 @@
 /*
- * Copyright (c) 2015-2016, Pradeeban Kathiravelu and others. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package edu.emory.bmi.datacafe.mongo;
+* Copyright (c) 2015-2016, Pradeeban Kathiravelu and others. All Rights Reserved.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+* http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+package edu.emory.bmi.datacafe.mongo.ext;
 
 import com.mongodb.client.FindIterable;
 import edu.emory.bmi.datacafe.constants.MongoConstants;
+import edu.emory.bmi.datacafe.mongo.MongoCollection;
+import edu.emory.bmi.datacafe.mongo.MongoConnector;
 import org.bson.Document;
 
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Methods for Mongo FindIterable.
- */
+* Methods for Mongo FindIterable.
+*/
+@Deprecated
 public class FindIterableConnector extends MongoConnector {
     /**
      * Gets the list of IDs
@@ -104,7 +107,4 @@ public class FindIterableConnector extends MongoConnector {
         }
         return iterableList;
     }
-
-
-
 }
