+++ modules/datacafe-server/src/main/java/edu/emory/bmi/datacafe/mongo/ext/package-info.java
@@ -1,0 +1,4 @@
+/*
+ * Extensions to the Mongo Integration, and Mongo Connectors.
+ */
+package edu.emory.bmi.datacafe.mongo.ext;
